Render notice if the user is logged in
Because of the complex and coupled signin logic, particulary related to social login, we must embed a passwordless login form in the existing signup block.

This step (user) provides an unspecified dependency [email]. Make sure to specify it in /signup/config/steps-pure.js, using the providesDependencies property.

Using simple promise to avoid catching every little error that Calypso throws
