+++ client/blocks/signup-form/style.scss
@@ -58,3 +58,15 @@
 	color: var( --color-text-inverted );
 	text-decoration: underline;
 }
+
+.signup-form__passwordless-form-wrapper {
+	.signup-form__terms-of-service-link {
+		margin: 4px 0;
+	}
+	.logged-out-form__footer {
+		margin-top: 0;
+	}
+	.validation-fieldset__validation-message {
+		min-height: auto;
+	}
+}
