+++ client/signup/config/steps-pure.js
@@ -524,16 +524,6 @@ export function generateSteps( {
 			providesDependencies: [ 'bearer_token', 'email', 'username' ],
 			unstorableDependencies: [ 'bearer_token' ],
 		},
-
-		'create-account': {
-			stepName: 'create-account',
-			providesToken: true,
-			providesDependencies: [ 'bearer_token', 'email', 'username' ],
-			unstorableDependencies: [ 'bearer_token' ],
-			props: {
-				isSocialSignupEnabled: config.isEnabled( 'signup/social' ),
-			},
-		},
 	};
 }
